Resolve compile issue with mojo refactor
